refactor(web): restructure the assignment form into five IA sections
Split the ~950-line DetailsSection monolith into five labeled sections (Details, Repository Setup, Autograding, Repository Features, Schedule), each a SectionCard with a per-section status badge derived from validateAssignmentForm (R2). Every gate now routes through deriveFormShape, and toSubmitValues/validation are unchanged, so the wire contract stays byte-identical (the characterization suite passes). Decouples the Feedback PR from autograding: it is now offered for any non-empty repo (Repository Setup), not just built-in autograding. Reserves the fork-mirror creation method and per-student Extensions as inert coming-soon affordances that write nothing. Slug/empty_repo/autograding-mode stay locked on edit; the read-only student and archived views are unaffected.
