+++ web/src/locales/en.json
@@ -1100,7 +1100,14 @@
     "goToAssignments": "Go to assignments",
     "form": {
       "detailsSection": "Details",
-      "settingsSection": "Settings",
+      "repositorySetupSection": "Repository Setup",
+      "repositoryFeaturesSection": "Repository Features",
+      "scheduleSection": "Schedule",
+      "sectionStatus": {
+        "error": "Needs attention",
+        "configured": "Configured",
+        "default": "Default"
+      },
       "optional": "optional",
       "name": "Assignment name",
       "namePlaceholder": "e.g., Loops Assignment",
@@ -1123,6 +1130,22 @@
       "typeGroup": "Group project",
       "maxGroupSize": "Max group size",
       "maxGroupSizeTitle": "Must be a whole number between {{min}} and {{max}}",
+      "customTriggersHeading": "Custom triggers",
+      "extensions": {
+        "label": "Per-student extensions (coming soon)",
+        "help": "Grant individual students or groups a later due date. Not yet available."
+      },
+      "creationMethod": {
+        "label": "Creation method",
+        "generate": {
+          "label": "Generate from template",
+          "help": "Create each repository from the template's default branch. This is the standard method."
+        },
+        "mirror": {
+          "label": "Mirror all branches and history (coming soon)",
+          "help": "A fork-like copy that mirrors every branch and the full commit history from the template. Not yet available."
+        }
+      },
       "feedbackPr": "Feedback pull request",
       "feedbackPrHelp": "Open a pull request per repo for inline review of each submission.",
       "feedbackPrEmptyRepoHelp": "Unavailable for empty repositories — there is no starter commit to diff against.",
@@ -1145,6 +1168,7 @@
       },
       "autograding": {
         "label": "Autograding",
+        "modeLabel": "Autograding mode",
         "emptyExplain": "Autograding is unavailable for empty repositories. Change the repository source in Repository Setup to use a template if you want autograding.",
         "lockedHelp": "The autograding mode can't be changed after creation: repositories students already accepted are not retrofitted with (or stripped of) the autograding workflow.",
         "choices": {
